Add optimistic search.

+++ src/get_search.h
@@ -20,7 +20,7 @@ class Search;
 extern unsigned int threads;
 extern fp_type cost_bound;
 extern unsigned int nblocks;
-extern float weight;
+extern double weight;
 
 Search *get_search(int argc, char *argv[]);
 vector<double> *parse_weights(char *str);
